Express delete route not returning error message
Struggled.  Returning a 204 code does not allow content to
follow (e.g. a crafted message).  Delete routing works and
fixed test but no clue how to get it to send an error msg
to Postman when deleting an already deleted object.

.send can accept the value of an object property but not the
object itself, apparently.  The key to that property will
show up in the res body, not the object parent.