Fix ENV loading error and move secret to .env
ENV var for JWT_SECRET was not loading to the environment.  Loaded
dotenv package into yarn and set a loading command in app.js
to set the variable into the Node Env.  When simply writing the
var into the environment use `source app.env`, it was not accessible
in the Node Env.  Also changed the file name to `.env` and set clear
in `.gitignore`.